fork at the right time the right way
Forking now is turned around.  After fork()
the forked command is the parent and
passfd becomes the child.

Why?  Because passfd goes away after passing the FDs,
but the forked command continues to exist usually.
And this way we do not need a waitpid() loop.

This way the parent can wait for the command to terminate,
not passfd.  For example this makes it possible to chain
ssh properly.